refactor(setu): 按功能模块拆分文件，日志降级为 Debug
- 拆分为 5 个文件：main / config / plugin / handler / media
- slog.Info 统一降级为 slog.Debug（非必要插件）
- 合并 handleHeisi / handleBaisi 为 handleSiImage，handleKkt 复用
